perf: optimize CPU reconstruction pipeline (FFT, Radon, FBP)
Reduce filtering and sinogram-construction overhead in the OpenMP backend:

- FFT now uses std::complex<float> instead of std::complex<double>: half
  the memory footprint, enables compiler SIMD, no precision loss for FBP.
- Add FFT::forwardInPlace / inverseInPlace working on caller-owned raw
  buffers. FilteredBackprojection and OpenMPBackend now reuse thread-local
  spectrum buffers instead of allocating std::vector<Complex> per angle.
- RadonTransform::forward parallelized via #pragma omp parallel for over
  angles (each thread writes to its own sinogram row, no races).
- schedule(dynamic) -> schedule(static) in FilteredBackprojection and
  OptimizedBackprojectionCPU where iteration cost is uniform.
- Extract duplicated AVX2 backprojection block in OptimizedBackprojectionCPU
  into backproject_row_avx2 helper used for both symmetric z-slices.
- Enable LTCG (/GL + /LTCG) for Release builds in CMakeLists.

+++ QtCTReconstruction/CTTypes.h
@@ -9,7 +9,7 @@
 
 namespace ct {
 
-using Complex = std::complex<double>;
+using Complex = std::complex<float>;
 
 // Замена std::vector<std::vector<float>> на плоский 1D массив с 2D-интерфейсом.
 struct Buffer2D {
